Editing the bedfile format
Requires some additional finishes to get the enhancer Grange upto scratch

Overlap promoters associated to enhancers with promoter Grange to get gene list

Overlap that with enhancers to assign gene targets
Overlap that with motifs in enhancers to assign targets
Need to put in blanks for motifs in enhancrs that are active but no known gene target. 

Thats my thought process.