Completed the pipe line (Paritally)
The first if statement where the package runs from nothing will now work and render everything appropriately. I need to copy and paste sections that need to be re-run when each variable is changed.

Additionally, i have added the script that i used to alter the cage expression file to an interaction bedpe file just for visualization
I have multiplied the correlation coefficients by 10 to be treated as counts for an approximation.