Fixed some typos and tested on different Computers
I have re-run the code on my mac to see if it will work as a stand alone: i have successfully gotten it to work, however, I noticed due to the limiations of my mac i was unable to visulize regions of the genome using the genome browser. Likely memory issues with my mac.