Optimized Speeds and Added exports
Consequence of observe function executing before anything loads and taking more than 2 minutes it i needed an alternative.

I subset the uniservse table for the entrez ids for the predicted UCSC gene ids and these are the respective inputs.

Secondly, I added 4 exports to the main tab so individuals may download each step of the genomic motif. 

also fixed a bug with the DNA string and the scoring system bugging out.